feat(core): use barel export everywhere - every folder should export the regular index barel not a modified version of it, for example Formatters, Filters, ... were all exporting their own set, these export were renamed as "formatters.index.ts", ... - this will allow to import any Editors, Filters, ... into Angular-Slickgrid
